Clean up file structure and add unary operators
