+++ src/value/attr_value/attr_value.rs
@@ -1,8 +1,9 @@
 use anyhow::*;
 use itertools::Itertools;
-use std::{collections::HashMap, iter::FromIterator};
+use serde::{Deserialize, Serialize};
+use std::{collections::HashMap, fmt, iter::FromIterator};
 
-use super::*;
+use super::super::*;
 
 /// A value assigned to an attribute in a widget.
 /// This can be a primitive String that contains any amount of variable
