+++ src/value/mod.rs
@@ -1,9 +1,7 @@
 use derive_more::*;
 use serde::{Deserialize, Serialize};
-use std::fmt;
 
 pub mod attr_value;
-pub mod attr_value_expr;
 pub mod coords;
 pub mod primitive;
 pub use attr_value::*;
